Pull down all from Triggerable to subtypes
