+++ src/main/java/org/javarosa/core/model/condition/Condition.java
@@ -20,12 +20,21 @@ import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import org.javarosa.core.model.QuickTriggerable;
 import org.javarosa.core.model.instance.FormInstance;
 import org.javarosa.core.model.instance.TreeElement;
 import org.javarosa.core.model.instance.TreeReference;
 import org.javarosa.core.util.externalizable.DeserializationException;
 import org.javarosa.core.util.externalizable.ExtUtil;
+import org.javarosa.core.util.externalizable.ExtWrapList;
+import org.javarosa.core.util.externalizable.ExtWrapTagged;
 import org.javarosa.core.util.externalizable.PrototypeFactory;
+import org.javarosa.debug.EvaluationResult;
+import org.javarosa.xpath.XPathException;
 
 public class Condition extends Triggerable {
     public static final int ACTION_NULL = 0;
@@ -124,7 +133,28 @@ public class Condition extends Triggerable {
             if (this == c)
                 return true;
 
-            return (this.trueAction == c.trueAction && this.falseAction == c.falseAction && super.equals(c));
+            boolean result = false;
+            if (c instanceof Triggerable) {
+                Triggerable t = (Triggerable) c;
+                if (this == t) {
+                    result = true;
+                } else if (this.expr.equals(t.expr)) {
+
+                    // The original logic did not make any sense --
+                    // the
+                    try {
+                        // resolved triggers should match...
+                        Set<TreeReference> Atriggers = this.getTriggers();
+                        Set<TreeReference> Btriggers = t.getTriggers();
+
+                        result = (Atriggers.size() == Btriggers.size()) &&
+                            Atriggers.containsAll(Btriggers);
+                    } catch (XPathException e) {
+                    }
+                }
+
+            }
+            return (this.trueAction == c.trueAction && this.falseAction == c.falseAction && result);
         } else {
             return false;
         }
@@ -132,18 +162,163 @@ public class Condition extends Triggerable {
 
     // region External serialization
 
+    @SuppressWarnings("unchecked")
     @Override
     public void readExternal(DataInputStream in, PrototypeFactory pf) throws IOException, DeserializationException {
-        super.readExternal(in, pf);
+        expr = (IConditionExpr) ExtUtil.read(in, new ExtWrapTagged(), pf);
+        contextRef = (TreeReference) ExtUtil.read(in, TreeReference.class, pf);
+        originalContextRef = (TreeReference) ExtUtil.read(in, TreeReference.class, pf);
+        List<TreeReference> tlist = (List<TreeReference>) ExtUtil.read(in, new ExtWrapList(TreeReference.class), pf);
+        targets = new ArrayList<>(tlist);
         trueAction = ExtUtil.readInt(in);
         falseAction = ExtUtil.readInt(in);
     }
 
     @Override
     public void writeExternal(DataOutputStream out) throws IOException {
-        super.writeExternal(out);
+        ExtUtil.write(out, new ExtWrapTagged(expr));
+        ExtUtil.write(out, contextRef);
+        ExtUtil.write(out, originalContextRef);
+        List<TreeReference> tlist = new ArrayList<>(targets);
+        ExtUtil.write(out, new ExtWrapList(tlist));
         ExtUtil.writeNumeric(out, trueAction);
         ExtUtil.writeNumeric(out, falseAction);
+    }
+
+    @Override
+    public void setImmediateCascades(Set<QuickTriggerable> cascades) {
+        immediateCascades = new HashSet<>(cascades);
+    }
+
+    @Override
+    public Set<QuickTriggerable> getImmediateCascades() {
+        return immediateCascades;
+    }
+
+    @Override
+    public TreeReference getContext() {
+        return contextRef;
+    }
+
+    @Override
+    public TreeReference getOriginalContext() {
+        return originalContextRef;
+    }
+
+    /**
+     * Dispatches all of the evaluation
+     */
+    @Override
+    public final List<EvaluationResult> apply(FormInstance mainInstance, EvaluationContext parentContext, TreeReference context) {
+        //The triggeringRoot is the highest level of actual data we can inquire about, but it _isn't_ necessarily the basis
+        //for the actual expressions, so we need genericize that ref against the current context
+        TreeReference ungenericised = originalContextRef.contextualize(context);
+        EvaluationContext ec = new EvaluationContext(parentContext, ungenericised);
+
+        Object result = eval(mainInstance, ec);
+
+        List<EvaluationResult> affectedNodes = new ArrayList<>(0);
+        for (TreeReference target : targets) {
+            TreeReference targetRef = target.contextualize(ec.getContextRef());
+            List<TreeReference> v = ec.expandReference(targetRef);
+
+            for (TreeReference affectedRef : v) {
+                apply(affectedRef, result, mainInstance);
+
+                affectedNodes.add(new EvaluationResult(affectedRef, result));
+            }
+        }
+
+        return affectedNodes;
+    }
+
+    public IConditionExpr getExpr() {
+        return expr;
+    }
+
+    public void addTarget(TreeReference target) {
+        if (targets.indexOf(target) == -1) {
+            targets.add(target);
+        }
+    }
+
+    @Override
+    public List<TreeReference> getTargets() {
+        return targets;
+    }
+
+    public Set<TreeReference> getTriggers() {
+        Set<TreeReference> relTriggers = expr.getTriggers(null);  /// should this be originalContextRef???
+        Set<TreeReference> absTriggers = new HashSet<>();
+        for (TreeReference r : relTriggers) {
+            absTriggers.add(r.anchor(originalContextRef));
+        }
+        return absTriggers;
+    }
+
+    Boolean evalPredicate(FormInstance model, EvaluationContext evalContext) {
+        try {
+            return expr.eval(model, evalContext);
+        } catch (XPathException e) {
+            e.setSource("Relevant expression for " + contextRef.toString(true));
+            throw e;
+        }
+    }
+
+    Object evalRaw(FormInstance model, EvaluationContext evalContext) {
+        try {
+            return expr.evalRaw(model, evalContext);
+        } catch (XPathException e) {
+            e.setSource("calculate expression for " + contextRef.toString(true));
+            throw e;
+        }
+    }
+
+    @Override
+    public void changeContextRefToIntersectWithTriggerable(Triggerable t) {
+        contextRef = contextRef.intersect(t.contextRef);
+    }
+
+    @Override
+    public TreeReference contextualizeContextRef(TreeReference anchorRef) {
+        // Contextualize the reference used by the triggerable against
+        // the anchor
+        return contextRef.contextualize(anchorRef);
+    }
+
+    @Override
+    public String toString() {
+        StringBuilder sb = new StringBuilder();
+        for (int i = 0; i < targets.size(); i++) {
+            sb.append(targets.get(i).toString());
+            if (i < targets.size() - 1)
+                sb.append(",");
+        }
+        return "trig[expr:" + expr.toString() + ";targets[" + sb.toString() + "]]";
+    }
+
+    /**
+     * Searches in the triggers of this Triggerable, trying to find the ones that are
+     * contained in the given list of contextualized refs.
+     *
+     * @param firedAnchorsMap a map of absolute refs
+     * @return a list of affected nodes.
+     */
+    @Override
+    public List<TreeReference> findAffectedTriggers(Map<TreeReference, List<TreeReference>> firedAnchorsMap) {
+        List<TreeReference> affectedTriggers = new ArrayList<>(0);
+
+        Set<TreeReference> triggers = this.getTriggers();
+        for (TreeReference trigger : triggers) {
+            List<TreeReference> firedAnchors = firedAnchorsMap.get(trigger.genericize());
+            if (firedAnchors == null) {
+                continue;
+            }
+
+            affectedTriggers.addAll(firedAnchors);
+        }
+
+        return affectedTriggers;
     }
 
     // endregion
